+++ src-gen/main/java/com/movisens/movisensgattlib/attributes/BatteryVoltageBuffered.java
@@ -80,7 +80,7 @@ public class BatteryVoltageBuffered extends AbstractBufferedAttribute<BatteryVol
 		
 		for (int i = 0; i < numValues; i++)
 		{
-			level[i] = new Double(bb.getUint8());
+			level[i] = new Double(bb.getUint16());
 		}
 	}
